Rename flows component to flows-page and update public API export
