+++ features/admin.flows.v1/public-api.ts
@@ -16,4 +16,4 @@
  * under the License.
  */
 
-export { default as FlowsPage } from "./pages/flows";
+export { default as FlowsPage } from "./pages/flows-page";
